Removing build-time warnings
Continuing to chip away at numerous varieties of build-time warnings and make minor code cleanup changes.

- Add a runtime ID to Stride.Samples.Tests so the compiler generates the same architecture as its dependencies.
- In Stride.Samples.Tests, remove link to AttachedChildProcessJob.cs as its not needed and results in a duplicate symbol warning.
- In AssetPropertGraph, actually use the isDisposed field.
- In several projects, us the released version of System.Collections.Immutable rather than the prerelease.
- In Stride.Core.Design, allow it to see the internal NativeLockFile compiled into Stride.Core.IO and avoid the duplicate symbol warning.
- Correct a few cases of sealed classes having protected members.
- As hygene, move ParticleBufferState to its own .cs
- In ParticleVertexBuilder, remove the unused indexStructSize
- Across multiple test projects, use the release version of Microsoft.NET.Test.Sdk rather than the prerelease version.
