+++ sources/core/Stride.Core.IO/Properties/AssemblyInfo.cs
@@ -1,10 +1,10 @@
-﻿// Copyright (c) .NET Foundation and Contributors (https://dotnetfoundation.org/ & https://stride3d.net) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
+// Copyright (c) .NET Foundation and Contributors (https://dotnetfoundation.org/ & https://stride3d.net) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
 // Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
-using System.Reflection;
 using System.Runtime.CompilerServices;
 
 #pragma warning disable 436 // Stride.PublicKeys is defined in multiple assemblies
 
+[assembly: InternalsVisibleTo("Stride.Core.Design" + Stride.PublicKeys.Default)]
 [assembly: InternalsVisibleTo("Stride.Core.IO.Serializers" + Stride.PublicKeys.Default)]
 [assembly: InternalsVisibleTo("Stride.Engine" + Stride.PublicKeys.Default)]
 [assembly: InternalsVisibleTo("Stride.Engine.Step1" + Stride.PublicKeys.Default)]
