+++ sources/engine/Stride.Particles/Spawners/SpawnerFromParent.cs
@@ -33,7 +33,7 @@ namespace Stride.Particles.Spawners
         /// Referenced parent emitter
         /// </summary>
         [DataMemberIgnore]
-        protected ParticleEmitter Parent;
+        private ParticleEmitter Parent;
 
         /// <summary>
         /// Referenced parent emitter's name
@@ -122,7 +122,7 @@ namespace Stride.Particles.Spawners
         /// Gets a field accessor to the parent emitter's spawn control field, if it exists
         /// </summary>
         /// <returns></returns>
-        protected ParticleFieldAccessor<ParticleChildrenAttribute> GetSpawnControlField()
+        private ParticleFieldAccessor<ParticleChildrenAttribute> GetSpawnControlField()
         {
             var groupIndex = (int)parentControlFlag;
             if (groupIndex >= ParticleFields.ChildrenFlags.Length)
