+++ sources/targets/Stride.UnitTests.targets
@@ -85,7 +85,7 @@
 
   <ItemGroup>
     <PackageReference Include="TeamCity.VSTest.TestAdapter" Version="1.0.32" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0-preview-20211130-02" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
   </ItemGroup>
 
   <Import Project="Sdk.targets" Sdk="Microsoft.NET.Sdk" />
